requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
huggingface-hub>=0.16.0
transformers>=4.33.0
accelerate>=0.21.0
diffusers>=0.21.0
compel>=2.0.0
omegaconf>=2.3.0
safetensors>=0.3.0
tqdm>=4.65.0
psutil>=5.9.0
gitpython>=3.1.0 